requirements.txt
python-dotenv
mailerlite